pyproject.toml
[tool.poetry]
name = "resume-matching-api"
version = "0.1.0"
description = "AI 기반 이력서-채용공고 매칭 서비스"
authors = ["JungWoochan"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"

# === 주요 프레임워크 & API ===
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
httpx = "^0.27.0"
pydantic = "^2.10.0"
python-multipart = "^0.0.6"

# === MongoDB ===
pymongo = "^4.11.0"
certifi = "^2024.2.2"

# === OpenAI 및 LLM 관련 ===
openai = "^1.68.0"
crewai = "^0.108.0"
langchain-openai = "^0.3.9"

# === 유틸리티 ===
python-dotenv = "^1.0.1"
pandas = "^2.2.3"

# === 미들웨어/서버 에러 핸들링 ===
starlette = "^0.46.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
